Thread control added to tests.
Fixes #29
Fixes robotpy/robotpy-wpilib#116

Because of the faked calls to Timer.delay any child threads
(used by PID loops, for example) would run their target functions
repeatedly with no delays.
This patch intercepts the delay call and uses a blocking call to wait
for a threading.Event. This makes the threads sleep until woken by
the main thread incrementing the time.
When a child thread requests a positive timestep that means it should
fire in the future, so it is made to sleep and woken if the time
advancement goes past this time. Threads that request negative time
should have fired in the previous increment of the timer, so they
are allowed to continue.

This has the nice effect that the child thread will fire the correct
number of times over the course of the test.